backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
duckdb>=0.9.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0

# Async
aiohttp>=3.9.0
asyncio>=3.4.3

# Utilities
PyYAML>=6.0
structlog>=23.2.0
python-dateutil>=2.8.2

# Reporting
weasyprint>=60.0
jinja2>=3.1.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
